requirements.txt
# FastAPI 웹서버
fastapi
uvicorn[standard]
requests

# Kafka 통신
kafka-python
pyyaml

# Triton Inference Server 통신
tritonclient[http]

# PyTorch 모델 생성 및 ONNX 변환
torch
onnx

# ML/통계 처리용
numpy
polars
scikit-learn

# 시각화
tqdm

# Data, Database
yfinance
duckdb